backend/requirements.txt
aiohttp==3.9.1
aiohttp-cors==0.7.0
httpx==0.26.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
cryptography==42.0.2
redis[asyncio]>=5.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
Pillow>=10.0.0
pywebpush>=2.0.0
aiosmtplib>=3.0.0
openai>=1.0.0
duckduckgo-search>=8.0.0
